// apps/sonora/lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      voice: {
        Row: {
          account_id: string
          created_at: string
          id: string
          is_default: boolean
          is_public: boolean
          name: string
          voice_id: string
        }
        Insert: {
          account_id?: string
          created_at?: string
          id?: string
          is_default?: boolean
          is_public?: boolean
          name: string
          voice_id: string
        }
        Update: {
          account_id?: string
          created_at?: string
          id?: string
          is_default?: boolean
          is_public?: boolean
          name?: string
          voice_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "voice_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "voice_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "user_account_workspace"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "voice_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "user_accounts"
            referencedColumns: ["id"]
          },
        ]
      }
    }
// ...
  }
// ...
}